unknown type uint replaced by uint8_t

+++ include/utils.h
@@ -7,7 +7,7 @@
 extern "C" {
 #endif
 
-char* generate_log_message_timestamp(uint log_level, uint32_t timestamp, char* log_message);
+char* generate_log_message_timestamp(uint8_t log_level, uint32_t timestamp, char* log_message);
 
 #ifdef __cplusplus
 }
